fix: datatypes

+++ y/datatypes.py
@@ -1,8 +1,8 @@
 
-class UsdValue(int):
+class UsdValue(float):
     def __init__(self, v) -> None:
         super().__init__()
 
-class UsdPrice(int):
+class UsdPrice(float):
     def __init__(self, v) -> None:
         super().__init__()
